requirements.txt
altair==5.2.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
extra-streamlit-components==0.1.70
fonttools==4.49.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
imbalanced-learn==0.11.0
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
leather==0.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
metrics==0.2.6
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pillow==10.2.0
plotly==5.20.0
protobuf==4.25.3
pyarrow==15.0.0
pydeck==0.8.1b0
pygal==3.0.4
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.31.1
streamlit-authenticator==0.3.1
streamlit-lottie==0.0.5
streamlit-metrics==0.1.0
tenacity==8.2.3
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
zipp==3.17.0
